[블로그개발_12] 소셜 로그인 붙이기 (spring-social-facebook) http://millky.com/@origoni/post/1170

+++ src/main/java/com/millky/blog/presentation/controller/PostController.java
@@ -5,7 +5,16 @@ import java.util.List;
 
 import javax.validation.Valid;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.social.connect.Connection;
+import org.springframework.social.connect.ConnectionData;
+import org.springframework.social.connect.ConnectionRepository;
+import org.springframework.social.facebook.api.Facebook;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -23,6 +32,9 @@ public class PostController {
 	@Autowired
 	private PostDao postDao;
 
+	@Autowired
+	private ConnectionRepository connectionRepository;
+
 	@RequestMapping(value = "/write", method = RequestMethod.GET)
 	public String form(Post post) {
 		return "form";
@@ -30,10 +42,15 @@ public class PostController {
 
 	@RequestMapping(value = "/write", method = RequestMethod.POST)
 	public String write(@Valid Post post, BindingResult bindingResult) {
+		User user = getConnect();
+
 		if (bindingResult.hasErrors()) {
 			return "form";
 		}
+
 		post.setRegDate(LocalDateTime.now());
+		post.setUserId(user.getProviderUserId());
+		post.setName(user.getDisplayName());
 		return "redirect:/post/" + postDao.save(post).getId();
 	}
 
@@ -41,6 +58,10 @@ public class PostController {
 	public String list(Model model) {
 		List<Post> postList = postDao.findAll();
 		model.addAttribute("postList", postList);
+
+		User user = getConnect();
+		model.addAttribute("user", user);
+
 		return "list";
 	}
 
@@ -48,27 +69,71 @@ public class PostController {
 	public String view(Model model, @PathVariable int id) {
 		Post post = postDao.findOne(id);
 		model.addAttribute("post", post);
+
+		User user = getConnect();
+		model.addAttribute("user", user);
+
 		return "post";
 	}
 
 	@RequestMapping("/{id}/delete")
 	public String delete(@PathVariable int id) {
-		postDao.delete(id);
+		User user = getConnect();
+
+		Post post = postDao.findOne(id);
+		if (post.getUserId().equals(user.getProviderUserId())) {
+			postDao.delete(id);
+		}
+
 		return "redirect:/post/list";
 	}
 
 	@RequestMapping(value = "/{id}/edit", method = RequestMethod.GET)
 	public String editor(Model model, @PathVariable int id) {
+		User user = getConnect();
+
 		Post post = postDao.findOne(id);
-		model.addAttribute("post", post);
+		if (post.getUserId().equals(user.getProviderUserId())) {
+			model.addAttribute("post", post);
+		}
+
 		return "form";
 	}
 
 	@RequestMapping(value = "/{id}/edit", method = RequestMethod.POST)
 	public String edit(@Valid Post post, BindingResult bindingResult) {
+		User user = getConnect();
+
 		if (bindingResult.hasErrors()) {
 			return "form";
 		}
-		return "redirect:/post/" + postDao.save(post).getId();
+
+		Post oldPost = postDao.findOne(post.getId());
+		if (oldPost.getUserId().equals(user.getProviderUserId())) {
+			oldPost.setTitle(post.getTitle());
+			oldPost.setSubtitle(post.getSubtitle());
+			oldPost.setContent(post.getContent());
+			return "redirect:/post/" + postDao.save(oldPost).getId();
+		}
+
+		return "form";
+	}
+
+	private User getConnect() {
+		Connection<Facebook> connection = connectionRepository.findPrimaryConnection(Facebook.class);
+		if (connection == null) {
+			return null;
+		}
+		ConnectionData data = connection.createData();
+		return new User(data.getProviderUserId(), data.getDisplayName());
+	}
+
+	@Getter
+	@Setter
+	@NoArgsConstructor
+	@AllArgsConstructor
+	public static class User {
+		String providerUserId;
+		String displayName;
 	}
 }
